dsgn impr + fix new wrks btn trg event twice

+++ src/components/Workspace.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import { createEventDispatcher, tick } from "svelte";
 	import Icon from "./Icon.svelte";
-	import { Picker } from "emoji-picker-element";
 	import EmojiPicker from "./EmojiPicker.svelte";
+	import { Key } from "ts-key-enum";
 
 	export let workspace: Ext.Workspace;
 	export let active = false;
@@ -20,6 +20,18 @@
 
 	export { classes as class };
 
+	function onKeyDown(e: KeyboardEvent) {
+		const { key } = e;
+
+		switch (key) {
+			case Key.Escape:
+				cancelEditing();
+				break;
+			default:
+				break;
+		}
+	}
+
 	function toggleEditMode() {
 		(async () => {
 			editMode = true;
@@ -34,11 +46,19 @@
 	}
 
 	function removeWorkspace() {
-		dispatch("removeWorkspace");
+		const removalConfirmed = confirm(
+			"This workspace will be removed. Continue?"
+		);
+
+		if (removalConfirmed) dispatch("removeWorkspace");
 	}
 
 	function switchWorkspace() {
 		dispatch("switchWorkspace");
+	}
+
+	function cancelEditing() {
+		editMode = false;
 	}
 
 	function openEmojiPicker(e: Event & { target: HTMLButtonElement }) {
@@ -70,25 +90,46 @@
 		class:active
 		class:selected
 		class="workspace grid gap-8 p-4 {editMode
-			? 'grid-cols-[max-content_1fr_max-content_max-content]'
+			? 'grid-cols-[max-content_1fr_max-content]'
 			: 'grid-cols-[1fr_max-content]'} justify-items-start items-center p-1 rounded-md {classes} focus-within:bg-neutral-200 focus-within:dark:bg-neutral-700 [&.active]:bg-[#5021ff] [&.active]:text-white"
 	>
 		{#if editMode}
-			<button on:click={openEmojiPicker} class="text-2xl">{iconValue}</button>
+			<button
+				title="pick emoji"
+				on:click={openEmojiPicker}
+				class="text-2xl rounded-full">{iconValue}</button
+			>
 			<input
-				class="bg-transparent border disabled:border-transparent"
+				class="bg-transparent border-b disabled:border-transparent outline-none"
 				{id}
 				type="text"
 				disabled={!editMode}
+				on:keydown={onKeyDown}
 				bind:this={nameInput}
 				bind:value={nameValue}
 			/>
-			<button on:click={removeWorkspace}
-				><Icon icon="remove" width={16} /></button
-			>
-			<button on:click={editWorkspace}>
-				<Icon icon="check" width={18} />
-			</button>
+			<div class="flex gap-2">
+				<button
+					title="remove"
+					on:click={removeWorkspace}
+					class="rounded-full outline-none focus:bg-white/25"
+					><Icon icon="bin" width={16} /></button
+				>
+				<button
+					title="apply"
+					on:click={editWorkspace}
+					class="rounded-full outline-none focus:bg-white/25"
+				>
+					<Icon icon="check" width={18} />
+				</button>
+				<button
+					title="cancel"
+					on:click={cancelEditing}
+					class="rounded-full outline-none focus:bg-white/25"
+				>
+					<Icon icon="cross" width={18} />
+				</button>
+			</div>
 		{:else}
 			<button
 				on:click={switchWorkspace}
@@ -100,7 +141,7 @@
 				<span class="{active ? 'font-bold' : ''} text-lg">{name}</span>
 				<!-- <span>({tabIds.join(",")})</span> -->
 			</button>
-			<button on:click={toggleEditMode}>
+			<button title="edit" on:click={toggleEditMode}>
 				<Icon icon="edit" width={14} />
 			</button>
 		{/if}
@@ -111,6 +152,10 @@
 	.workspace {
 		@media screen and (width < 260px) {
 			@apply aspect-square grid-cols-1 h-12 p-0 justify-items-center justify-self-center;
+
+			button:first-of-type span {
+				@apply w-full;
+			}
 
 			button span:nth-of-type(n + 2) {
 				@apply hidden;
